Fusion QuicFizzFactory into FizzCryptoFactory
