+++ quic/server/handshake/ServerHandshake.cpp
@@ -35,7 +35,9 @@ void ServerHandshake::initialize(
     std::unique_ptr<fizz::server::AppTokenValidator> validator) {
   executor_ = executor;
   auto ctx = std::make_shared<fizz::server::FizzServerContext>(*context);
-  ctx->setFactory(std::make_shared<QuicFizzFactory>());
+  auto cryptoFactory = std::make_shared<FizzCryptoFactory>();
+  ctx->setFactory(cryptoFactory);
+  cryptoFactory_ = std::move(cryptoFactory);
   ctx->setSupportedCiphers({{fizz::CipherSuite::TLS_AES_128_GCM_SHA256}});
   ctx->setVersionFallbackEnabled(false);
   // Since Draft-17, client won't sent EOED
@@ -396,9 +398,7 @@ void ServerHandshake::ActionMoveVisitor::operator()(
       folly::range(secretAvailable.secret.secret),
       kQuicKeyLabel,
       kQuicIVLabel);
-  QuicFizzFactory fizzFactory;
-  FizzCryptoFactory cryptoFactory(&fizzFactory);
-  auto headerCipher = cryptoFactory.makePacketNumberCipher(
+  auto headerCipher = server_.cryptoFactory_->makePacketNumberCipher(
       folly::range(secretAvailable.secret.secret));
   folly::variant_match(
       secretAvailable.secret.type,
